requirements.txt
kafka-python==2.0.2
psutil==5.9.8
psycopg2-binary==2.9.9
pandas==2.2.2
streamlit==1.36.0
numpy==1.26.4
scikit-learn==1.4.2
